+++ package.json
@@ -32,8 +32,7 @@
     "react-jss": "^10.5.0",
     "react-pose": "^4.0.10",
     "react-pose-text": "^3.1.0",
-    "rxjs": "^6.6.2",
-    "typestyle": "^2.1.0"
+    "rxjs": "^6.6.2"
   },
   "devDependencies": {
     "@babel/core": "^7.11.4",
