+++ src/components/Button/style.ts
@@ -115,5 +115,42 @@ export const style = createUseStyles(
   { theming, name: 'sha-el-button' },
 );
 
+export const buttonGroupStyle = createUseStyles(
+  {
+    buttonGroup: {
+      '& button': {
+        borderRadius: '0',
+        '&:first-child': {
+          borderRadius: '5px 0 0 5px',
+        },
+        '&:last-child': {
+          borderRadius: '0 5px 5px 0',
+        },
+      },
+      '& a': {
+        boxSizing: 'border-box',
+        borderRadius: '0',
+        '&:first-child': {
+          borderRadius: '5px 0 0 5px',
+        },
+        '&:last-child': {
+          borderRadius: '0 5px 5px 0',
+        },
+      },
+      '& *:first-child': {
+        '& button': {
+          borderRadius: '5px 0 0 5px !important',
+        },
+      },
+      '& *:last-child': {
+        '& button': {
+          borderRadius: '0 5px 5px 0 !important',
+        },
+      },
+    },
+  },
+  { theming, name: 'sha-el-button-group' },
+);
+
 export type sizeTypes = 'default' | 'big' | 'small';
 export type shapeTypes = 'default' | 'circle';
