+++ src/components/Grid/Col.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
-import { createUseStyles } from 'react-jss';
 import { classes } from '../../helpers';
+import { colStyle } from './style';
 
 export const Col: React.FC<ColProps> = React.forwardRef((props, ref) => {
   const {
@@ -20,59 +20,6 @@ export const Col: React.FC<ColProps> = React.forwardRef((props, ref) => {
     </div>
   );
 });
-
-const colStyle = createUseStyles(
-  {
-    col: (props) => ({
-      maxWidth: !props.flex && 100 / (24 / props.span) + '%',
-      flexBasis: !props.flex && 100 / (24 / props.span) + '%',
-      boxSizing: 'border-box',
-      flex: props.flex || '0 0 auto',
-      marginLeft: props.offset && typeof props.offset === 'number' && 100 / (24 / props.offset) + '%',
-      alignSelf: props.alignSelf,
-    }),
-    '@media (max-width: 576px)': {
-      col: (props) => ({
-        maxWidth: !props.flex && 100 / (24 / props.spanXs) + '%',
-        flexBasis: !props.flex && 100 / (24 / props.spanXs) + '%',
-        marginLeft: props.offset && typeof props.offset === 'object' && 100 / (24 / props.offset.xs) + '%',
-      }),
-    },
-
-    '@media (min-width: 576px)': {
-      col: (props) => ({
-        maxWidth: !props.flex && 100 / (24 / props.spanSm) + '%',
-        flexBasis: !props.flex && 100 / (24 / props.spanSm) + '%',
-        marginLeft: props.offset && typeof props.offset === 'object' && 100 / (24 / props.offset.sm) + '%',
-      }),
-    },
-
-    '@media (min-width: 768px)': {
-      col: (props) => ({
-        maxWidth: !props.flex && 100 / (24 / props.spanMd) + '%',
-        flexBasis: !props.flex && 100 / (24 / props.spanMd) + '%',
-        marginLeft: props.offset && typeof props.offset === 'object' && 100 / (24 / props.offset.md) + '%',
-      }),
-    },
-
-    '@media (min-width: 992px)': {
-      col: (props) => ({
-        maxWidth: !props.flex && 100 / (24 / props.spanLg) + '%',
-        flexBasis: !props.flex && 100 / (24 / props.spanLg) + '%',
-        marginLeft: props.offset && typeof props.offset === 'object' && 100 / (24 / props.offset.lg) + '%',
-      }),
-    },
-
-    '@media (min-width: 1200px)': {
-      col: (props) => ({
-        maxWidth: !props.flex && 100 / (24 / props.spanXl) + '%',
-        flexBasis: !props.flex && 100 / (24 / props.spanXl) + '%',
-        marginLeft: props.offset && typeof props.offset === 'object' && 100 / (24 / props.offset.xl) + '%',
-      }),
-    },
-  },
-  { name: 'sha-el-col' },
-);
 
 export interface ColProps extends React.DetailedHTMLProps<React.HTMLAttributes<HTMLDivElement>, HTMLDivElement> {
   style?: React.CSSProperties;
