+++ src/components/Grid/Row.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
-import { createUseStyles } from 'react-jss';
 import { classes } from '../../helpers';
+import { flexPosition, rowStyle } from './style';
 
 export const Row: React.FC<RowProps> = (props) => {
   const { className } = props;
@@ -20,26 +20,6 @@ Row.defaultProps = {
   gutter: ['10px 0', '.5rem'],
   wrap: 'wrap',
 };
-
-const rowStyle = createUseStyles({
-  container: {
-    boxSizing: 'border-box',
-    display: 'flex',
-    flexDirection: 'row',
-  },
-});
-
-const flexPosition = createUseStyles({
-  container: (props) => ({
-    justifyContent: props.justifyContent,
-    alignItems: props.alignItems,
-    padding: props.gutter[0],
-    flexWrap: props.wrap,
-    '& > .sha-el-col': {
-      padding: props.gutter[1],
-    },
-  }),
-});
 
 export interface RowProps {
   children: React.ReactNode;
