+++ src/components/InfiniteScroll/style.ts
@@ -1,19 +1,22 @@
 import { createUseStyles } from 'react-jss';
 
-export const style = createUseStyles({
-  container: ({ height }) => ({
-    height: height || 'auto',
-    overflow: height ? 'auto' : 'hidden',
-    position: 'relative',
-  }),
-  loadingIndicator: {
-    marginBottom: '20px',
-    position: 'relative',
-    margin: 'auto',
-    left: 0,
-    right: 0,
+export const style = createUseStyles(
+  {
+    container: ({ height }) => ({
+      height: height || 'auto',
+      overflow: height ? 'auto' : 'hidden',
+      position: 'relative',
+    }),
+    loadingIndicator: {
+      marginBottom: '20px',
+      position: 'relative',
+      margin: 'auto',
+      left: 0,
+      right: 0,
+    },
+    loadingContainer: {
+      position: 'relative',
+    },
   },
-  loadingContainer: {
-    position: 'relative',
-  },
-});
+  { name: 'sha-el-infinite-scroll' },
+);
