+++ src/components/Radio/RadioButton.tsx
@@ -1,10 +1,10 @@
 import * as React from 'react';
-import { stylesheet } from 'typestyle';
 
 import { Button } from '../Button';
 import { RadioProps } from './Radio';
-import { ThemeConsumer, Theme } from '../Theme/Theme';
+import { useTheme } from '../Theme/Theme';
 import { nestedAccess } from '../../helpers';
+import { radioButtonStyle as style } from './style';
 
 export const RadioButton: React.FunctionComponent<RadioProps> = (props) => {
   const { label, className, error, checked, disabled } = props;
@@ -17,39 +17,19 @@ export const RadioButton: React.FunctionComponent<RadioProps> = (props) => {
     }
   };
 
+  const theme = useTheme();
+  const css = style(theme);
   return (
-    <ThemeConsumer>
-      {(theme) => {
-        const css = style(props.block, theme);
-        return (
-          <div style={{ display: 'inline' }} onClick={() => onContainerClick()}>
-            <input className={css.radio} ref={input} type="radio" {...props} />
-            <Button
-              disabled={disabled}
-              type={checked || nestedAccess(input.current, 'checked') ? 'primary' : 'default'}
-            >
-              {label}
-            </Button>
-            {error && (
-              <div key="error" className={`${css.errorStyle} ${className || ''}`}>
-                {error}
-              </div>
-            )}
-          </div>
-        );
-      }}
-    </ThemeConsumer>
+    <div style={{ display: 'inline' }} onClick={() => onContainerClick()}>
+      <input className={css.radio} ref={input} type="radio" {...props} />
+      <Button disabled={disabled} type={checked || nestedAccess(input.current, 'checked') ? 'primary' : 'default'}>
+        {label}
+      </Button>
+      {error && (
+        <div key="error" className={`${css.errorStyle} ${className || ''}`}>
+          {error}
+        </div>
+      )}
+    </div>
   );
-};
-
-const style = (block: boolean, theme: Theme) => {
-  return stylesheet({
-    errorStyle: {
-      fontSize: '14px',
-      color: theme.error,
-    },
-    radio: {
-      display: 'none',
-    },
-  });
 };
